+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## []
+### Changed
+- Changed User Interface using icons
+- Changed library to interface with SSD1306 display
+
 ### Removed
 - Removed custom PID controller and use default one included in lib_coffee_machine